+++ changes.md
@@ -1,1 +1,2 @@
 * Cache configuration so white/black lists are not fetched all the time (#56)
+* Add a configuration to whitelist the blocks that can be adjacent to a trunk. By default (empty) this won't apply any restrictions (#54)